+++ tests/units/array/test_array.py
@@ -5,7 +5,6 @@ import pytest
 import torch
 
 from docarray import BaseDocument, DocumentArray
-from docarray.array import DocumentArrayStacked
 from docarray.typing import NdArray, TorchTensor
 
 
@@ -254,26 +253,3 @@ def test_get_from_slice():
     assert len(texts) == 5
     for i, text in enumerate(texts):
         assert text == f'hello{i*2}'
-
-
-def test_get_from_slice_stacked():
-    class Doc(BaseDocument):
-        text: str
-        tensor: NdArray
-
-    N = 10
-
-    da = DocumentArray[Doc](
-        (Doc(text=f'hello{i}', tensor=np.zeros((3, 224, 224))) for i in range(N))
-    ).stack()
-
-    da_sliced = da[0:10:2]
-    assert isinstance(da_sliced, DocumentArrayStacked)
-
-    tensors = da_sliced.tensor
-    assert tensors.shape == (5, 3, 224, 224)
-
-    texts = da_sliced.text
-    assert len(texts) == 5
-    for i, text in enumerate(texts):
-        assert text == f'hello{i * 2}'
